Rename `comp.*` to `compression.*`
