+++ cpp/src/io/orc/orc_gpu.hpp
@@ -16,7 +16,7 @@
 
 #pragma once
 
-#include "io/comp/comp.hpp"
+#include "io/comp/compression.hpp"
 #include "io/statistics/statistics.cuh"
 #include "orc.hpp"
