+++ cpp/tests/large_strings/json_tests.cu
@@ -15,7 +15,7 @@
  */
 
 #include "../io/json/json_utils.cuh"
-#include "io/comp/comp.hpp"
+#include "io/comp/compression.hpp"
 #include "io/comp/io_uncomp.hpp"
 #include "large_strings_fixture.hpp"
